Refactor CodeMirrorInput to its own file and use it for the value input as well
