+++ studio/src/components/Timeline/DetailsList/CodeMirrorEditor/index.ts
@@ -1,0 +1,7 @@
+export {
+  CodeMirrorTypescriptEditor,
+  CodeMirrorJsonEditor,
+  CodeMirrorSqlEditor,
+} from "./CodeMirrorEditor";
+
+export { CodeMirrorInput } from "./CodeMirrorInput";
